Extract the common code between the Part 1 ont the Part 2 of the day 9 in a library.
