+++ 2015/Day9/src/input_parser.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "types.hpp"
+
+std::pair<City, City> getCitiesFromInstruction (const std::string_view instruction);
+
+Distance getDistanceFromInstruction (const std::string& instruction);
